Remove ring library, refactor to use `crypto` lib
