+++ src/contract/error.rs
@@ -22,26 +22,26 @@ impl fmt::Display for Error {
 }
 
 impl From<ethabi::Error> for Error {
-    fn from(err: ethabi::Error) -> Self {
-        Error::InvalidContract("".to_string())
+    fn from(_: ethabi::Error) -> Self {
+        Error::InvalidContract("ethabi error".to_string())
     }
 }
 
 impl From<ethabi::spec::Error> for Error {
-    fn from(err: ethabi::spec::Error) -> Self {
-        Error::InvalidContract("".to_string())
+    fn from(_: ethabi::spec::Error) -> Self {
+        Error::InvalidContract("ethabi spec error".to_string())
     }
 }
 
 impl From<ethabi::spec::param_type::Error> for Error {
-    fn from(err: ethabi::spec::param_type::Error) -> Self {
-        Error::InvalidContract("".to_string())
+    fn from(_: ethabi::spec::param_type::Error) -> Self {
+        Error::InvalidContract("ethabi param error".to_string())
     }
 }
 
 impl From<ethabi::token::Error> for Error {
-    fn from(err: ethabi::token::Error) -> Self {
-        Error::InvalidContract("".to_string())
+    fn from(_: ethabi::token::Error) -> Self {
+        Error::InvalidContract("ethabi token error".to_string())
     }
 }
