+++ src/hdwallet/bip32.rs
@@ -1,5 +1,5 @@
 //! # Module to generate private key from HD path
-//! according to the [BIP32]()
+//! according to the [BIP32](https://github.com/bitcoin/bips/blob/master/bip-0032.mediawiki)
 //!
 
 use super::error::Error;
@@ -8,9 +8,35 @@ use bitcoin::util::bip32::ChildNumber::{self, Hardened, Normal};
 use bitcoin::util::bip32::ExtendedPrivKey;
 use core::{PRIVATE_KEY_BYTES, PrivateKey};
 use regex::Regex;
+use hdwallet::DERIVATION_INDEX_SIZE;
 use secp256k1::Secp256k1;
 use std::ops;
+use util::to_bytes;
 
+
+pub const ETC_DERIVATION_PATH: [u8; 21] = [
+    5,
+    0x80,
+    0,
+    0,
+    44,
+    0x80,
+    0,
+    0,
+    60,
+    0x80,
+    0x02,
+    0x73,
+    0xd0,
+    0x80,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+]; // 44'/60'/160720'/0'/0
 
 lazy_static! {
     static ref HD_PATH_RE: Regex = Regex::new(r#"^m/{1}[^0-9'/]*"#).unwrap();
@@ -34,7 +60,7 @@ impl HDPath {
         let mut res: Vec<ChildNumber> = vec![];
 
         if !HD_PATH_RE.is_match(path) {
-            return Err(Error::KeyGenerationError(
+            return Err(Error::HDWalletError(
                 "Invalid HD path format".to_string(),
             ));
         }
@@ -58,7 +84,7 @@ impl HDPath {
                     }
                 }
                 Err(e) => {
-                    return Err(Error::KeyGenerationError(
+                    return Err(Error::HDWalletError(
                         format!("Invalid HD path child index: {}", e.to_string()),
                     ))
                 }
@@ -91,6 +117,63 @@ pub fn generate_key(path: &HDPath, seed: &[u8]) -> Result<PrivateKey, Error> {
     let key = PrivateKey::try_from(&sk.secret_key[0..PRIVATE_KEY_BYTES])?;
 
     Ok(key)
+}
+
+/// Parse HD path into byte array
+///
+/// # Arguments:
+///
+/// * hd_str - path string
+///
+pub fn path_to_arr(hd_str: &str) -> Result<Vec<u8>, Error> {
+    if !HD_PATH_RE.is_match(hd_str) {
+        return Err(Error::HDWalletError(
+            format!("Invalid `hd_path` format: {}", hd_str),
+        ));
+    }
+
+    let (_, p) = hd_str.split_at(2);
+    let mut buf = Vec::new();
+    {
+        let mut parse = |s: &str| {
+            let mut str = s.to_string();
+            let mut v: u64 = 0;
+
+            if str.ends_with('\'') {
+                v += 0x8000_0000;
+                str.remove(s.len() - 1);
+            }
+            match str.parse::<u64>() {
+                Ok(d) => v += d,
+                Err(_) => {
+                    return Err(Error::HDWalletError(
+                        format!("Invalid index: {}", hd_str),
+                    ))
+                }
+            }
+            buf.extend(to_bytes(v, 4));
+            Ok(())
+        };
+
+        for val in p.split('/') {
+            parse(val)?;
+        }
+    }
+
+    Ok(buf)
+}
+
+/// Parse HD path into byte array
+/// prefixed with count of derivation indexes
+pub fn to_prefixed_path(hd_str: &str) -> Result<Vec<u8>, Error> {
+    let v = path_to_arr(hd_str)?;
+    let count = (v.len() / DERIVATION_INDEX_SIZE) as u8;
+    let mut buf = Vec::with_capacity(v.len() + 1);
+
+    buf.push(count);
+    buf.extend(v);
+
+    Ok(buf)
 }
