+++ src/lib.rs
@@ -35,7 +35,6 @@ extern crate hex;
 extern crate hidapi;
 extern crate emerald_rocksdb as rocksdb;
 extern crate num;
-extern crate ring;
 extern crate bitcoin;
 
 mod core;
