+++ src/rpc/serves.rs
@@ -3,7 +3,8 @@ use super::StorageController;
 use super::serialize::RPCTransaction;
 use contract::Contract;
 use core::{Address, Transaction};
-use hdwallet::{WManager, to_prefixed_path};
+use hdwallet::WManager;
+use hdwallet::bip32::to_prefixed_path;
 use jsonrpc_core::{Params, Value};
 use keystore::{CryptoType, Kdf, KdfDepthLevel, KeyFile, PBKDF2_KDF_NAME, os_random};
 use mnemonic::{self, ENTROPY_BYTE_LENGTH, HDPath, Language, Mnemonic, gen_entropy};
